第十二次上传： 1. 修正了地图的规格大小 2. 添加了Hpbar类，并调整了Monster和Headquarterd的show()方法，现在可以显示怪物和基地的血量

+++ Tower_Defence/Icon.h
@@ -11,16 +11,16 @@ class Icon
         static map<string, Icon> Icon_Set;
         static Icon findIcon(string type);
 
-        Icon(string type, int x, int y, double w, double h);
-        int getScrX() const {return _scrX;};
-        int getScrY() const {return _scrY;};
+        Icon(string type, double x, double y, double w, double h);
+        double getScrX() const {return _scrX;};
+        double getScrY() const {return _scrY;};
         double getHeight() const {return _height;};
         double getWidth() const {return _width;};
         string getType() const {return this->_typeName;};
 
     protected:
         string _typeName;
-        int _scrX, _scrY;
+        double _scrX, _scrY;
         double _height, _width;
 };
